ADD: avatar router.
