+++ src/server/index.ts
@@ -1,22 +1,19 @@
-import path from "path";
 import fs from "fs";
 
-import { XUtils } from "@vex-chat/crypto";
 import { XTypes } from "@vex-chat/types";
 import cors from "cors";
 import express from "express";
 import expressWs from "express-ws";
-import FileType from "file-type";
 
 import helmet from "helmet";
 import morgan from "morgan";
-import multer from "multer";
-import nacl from "tweetnacl";
 import winston from "winston";
 
 import { Database } from "../Database";
+
 import { getUserRouter } from "./user";
 import { getFileRouter } from "./file";
+import { getAvatarRouter } from "./avatar";
 
 // expiry of regkeys
 export const EXPIRY_TIME = 1000 * 60 * 5;
@@ -39,7 +36,9 @@ export const initApp = (
     // INIT ROUTERS
     const userRouter = getUserRouter(db, log, tokenValidator);
     const fileRouter = getFileRouter(db, log);
+    const avatarRouter = getAvatarRouter(db, log);
 
+    // MIDDLEWARE
     api.use(express.json({ limit: "20mb" }));
     api.use(helmet());
 
@@ -49,6 +48,7 @@ export const initApp = (
 
     api.use(cors());
 
+    // SIMPLE RESOURCES
     api.get("/server/:id", async (req, res) => {
         const server = await db.retrieveServer(req.params.id);
 
@@ -83,70 +83,12 @@ export const initApp = (
         res.send({ canary: process.env.CANARY });
     });
 
+    // COMPLEX RESOURCES
     api.use("/user", userRouter);
 
-    // file
     api.use("/file", fileRouter);
 
-    // avatar
-    api.get("/avatar/:userID", async (req, res) => {
-        fs.readFile(
-            path.resolve("./avatars/" + req.params.userID),
-            undefined,
-            async (err, file) => {
-                if (err) {
-                    log.error("error reading file");
-                    log.error(err);
-                    res.sendStatus(404);
-                } else {
-                    const typeDetails = await FileType.fromBuffer(file);
-                    if (typeDetails) {
-                        res.set("Content-type", typeDetails.mime);
-                    }
-                    res.send(file);
-                }
-            }
-        );
-    });
-
-    api.post("/avatar/:userID", multer().single("avatar"), async (req, res) => {
-        const payload: XTypes.HTTP.IFilePayload = req.body;
-        const userEntry = await db.retrieveUser(req.params.userID);
-
-        if (!userEntry) {
-            res.sendStatus(404);
-            return;
-        }
-
-        const devices = await db.retrieveUserDeviceList(req.params.userID);
-
-        let token: Uint8Array | null = null;
-        for (const device of devices) {
-            const verified = nacl.sign.open(
-                XUtils.decodeHex(payload.signed),
-                XUtils.decodeHex(device.signKey)
-            );
-            if (verified) {
-                token = verified;
-            }
-        }
-        if (!token) {
-            log.warn("Bad signature on token.");
-            res.sendStatus(401);
-            return;
-        }
-
-        try {
-            // write the file to disk
-            fs.writeFile("avatars/" + userEntry.userID, req.file.buffer, () => {
-                log.info("Wrote new avatar " + userEntry.userID);
-            });
-            res.sendStatus(200);
-        } catch (err) {
-            log.warn(err);
-            res.sendStatus(500);
-        }
-    });
+    api.use("/avatar", avatarRouter);
 };
 
 /**
